added lead functionality, added `Leashable` trait and moved leash code outside `mob` crate
