+++ steel-core/src/behavior/items/mod.rs
@@ -33,6 +33,7 @@ mod throwable_potion;
 mod tipped_arrow;
 
 mod flint_and_steel;
+mod lead;
 
 pub use air::AirItem;
 pub use axe::AxeItem;
@@ -51,6 +52,7 @@ pub use flint_and_steel::{FireChargeItem, FlintAndSteelItem};
 pub use food_on_a_stick::FoodOnAStickItem;
 pub use hoe::HoeItem;
 pub use honeycomb::HoneycombItem;
+pub use lead::LeadItem;
 pub use mace::MaceItem;
 pub use name_tag::NameTagItem;
 pub use player_head::PlayerHeadItem;
