+++ steel-core/src/entity/entities/mobs/passive/pig/tests/persistence.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::entity::leash::{LeashAttachment, Leashable};
 
 #[test]
 fn pig_saves_vanilla_mob_age_and_variant_data() {
