+++ steel-core/src/entity/leash.rs
@@ -1,0 +1,493 @@
+use std::f32::consts::PI;
+use std::sync::Arc;
+
+use crate::entity::entities::LeashFenceKnotEntity;
+use crate::entity::{Entity, Mob, SharedEntity, WeakEntity};
+use glam::DVec3;
+use simdnbt::borrow::NbtCompound as BorrowedNbtCompoundView;
+use simdnbt::owned::{NbtCompound, NbtTag};
+use steel_protocol::packets::game::SoundSource;
+use steel_registry::blocks::block_state_ext::BlockStateExt;
+use steel_registry::item_stack::ItemStack;
+use steel_registry::vanilla_game_rules::ENTITY_DROPS;
+use steel_registry::{sound_events, vanilla_items};
+use steel_utils::locks::SyncMutex;
+use steel_utils::{BlockPos, Downcast as _, UuidExt as _};
+use uuid::Uuid;
+
+pub(super) const LEASH_SNAP_DISTANCE: f64 = 12.0;
+pub(super) const LEASH_ELASTIC_DISTANCE: f64 = 6.0;
+pub(super) const LEASH_AXIS_SPECIFIC_ELASTICITY: DVec3 = DVec3::new(0.8, 0.2, 0.8);
+pub(super) const LEASH_SPRING_DAMPENING: f64 = 0.7;
+pub(super) const LEASH_TORSIONAL_ELASTICITY: f64 = 10.0;
+pub(super) const LEASH_STIFFNESS: f64 = 0.11;
+pub(super) const ENTITY_LEASH_ATTACHMENT_POINT: DVec3 = DVec3::new(0.0, 0.5, 0.5);
+pub(super) const LEASHER_ATTACHMENT_POINT: DVec3 = DVec3::new(0.0, 0.5, 0.0);
+pub(super) const DELAYED_LEASH_DROP_TICKS: i32 = 100;
+
+/// Vanilla-shaped behavior shared by entities that extend `Leashable`.
+pub trait Leashable: Entity {
+    fn leash_data(&self) -> &SyncMutex<Option<LeashData>>;
+
+    fn is_leashed(&self) -> bool {
+        self.leash_holder().is_some()
+    }
+
+    fn may_be_leashed(&self) -> bool {
+        self.leash_data().lock().is_some()
+    }
+
+    fn leash_holder(&self) -> Option<SharedEntity> {
+        self.leash_data()
+            .lock()
+            .as_ref()
+            .and_then(LeashData::holder)
+    }
+
+    fn leash_attachment(&self) -> Option<LeashAttachment> {
+        self.leash_data()
+            .lock()
+            .as_ref()
+            .map(LeashData::saved_attachment)
+    }
+
+    fn set_delayed_leash_attachment(&self, attachment: LeashAttachment) {
+        *self.leash_data().lock() = Some(LeashData::from_delayed_attachment(attachment));
+    }
+
+    fn can_be_leashed(&self) -> bool {
+        // TODO: Return false for enemy mobs once hostile mob foundations exist.
+        true
+    }
+
+    fn leash_distance_to(&self, holder: &dyn Entity) -> f64 {
+        leash_bounding_box_center(self.as_entity_event_source())
+            .distance(leash_bounding_box_center(holder))
+    }
+
+    fn leash_snap_distance(&self) -> f64 {
+        LEASH_SNAP_DISTANCE
+    }
+
+    fn leash_elastic_distance(&self) -> f64 {
+        LEASH_ELASTIC_DISTANCE
+    }
+
+    fn when_leashed_to(&self, holder: &dyn Entity) {
+        holder.notify_leash_holder(self.as_entity_event_source());
+    }
+
+    fn leash_too_far_behaviour(&self) {
+        self.drop_leash();
+    }
+
+    fn on_elastic_leash_pull(&self) {
+        self.check_fall_distance_accumulation();
+    }
+
+    fn close_range_leash_behaviour(&self, _holder: &dyn Entity) {}
+
+    fn check_elastic_interactions(&self, holder: &dyn Entity) -> bool {
+        let Some(wrench) = compute_elastic_interaction(
+            self.as_entity_event_source(),
+            holder,
+            self.leash_elastic_distance(),
+        ) else {
+            return false;
+        };
+
+        {
+            let mut leash_data = self.leash_data().lock();
+            let Some(leash_data) = leash_data.as_mut() else {
+                return false;
+            };
+            leash_data.angular_momentum += LEASH_TORSIONAL_ELASTICITY * wrench.torque;
+        }
+
+        let relative_velocity_to_leasher =
+            leash_holder_movement(holder) - leash_holder_movement(self.as_entity_event_source());
+        self.push_impulse(
+            axis_specific_leash_elasticity(wrench.force)
+                + relative_velocity_to_leasher * LEASH_STIFFNESS,
+        );
+        true
+    }
+
+    fn apply_leash_angular_momentum(&self) -> bool {
+        let angular_friction = self.leash_angular_friction();
+        let angular_momentum = {
+            let mut leash_data = self.leash_data().lock();
+            let Some(leash_data) = leash_data.as_mut() else {
+                return false;
+            };
+            let angular_momentum = leash_data.angular_momentum;
+            leash_data.angular_momentum *= angular_friction;
+            angular_momentum
+        };
+        self.rotate_by_leash_angular_momentum(angular_momentum);
+        true
+    }
+
+    fn rotate_by_leash_angular_momentum(&self, angular_momentum: f64) {
+        let (yaw, pitch) = self.rotation();
+        self.set_rotation((yaw - angular_momentum as f32, pitch));
+    }
+
+    fn leash_angular_momentum(&self) -> Option<f64> {
+        self.leash_data()
+            .lock()
+            .as_ref()
+            .map(|leash_data| leash_data.angular_momentum)
+    }
+
+    fn leash_angular_friction(&self) -> f64 {
+        if self.on_ground() {
+            let Some(world) = self.level() else {
+                return 0.91;
+            };
+            let Some(pos) = self.block_pos_below_that_affects_movement() else {
+                return 0.91;
+            };
+            return f64::from(world.get_block_state(pos).get_block().config.friction * 0.91);
+        }
+
+        if self.is_in_water() || self.is_in_lava() {
+            return 0.8;
+        }
+
+        0.91
+    }
+
+    fn can_have_a_leash_attached_to(&self, holder: &dyn Entity) -> bool {
+        self.id() != holder.id()
+            && self.leash_distance_to(holder) <= self.leash_snap_distance()
+            && self.can_be_leashed()
+    }
+
+    fn set_leashed_to(&self, holder: &SharedEntity) -> bool {
+        if self.id() == holder.id() {
+            return false;
+        }
+
+        let old_holder = self.leash_holder();
+        {
+            let mut leash_data = self.leash_data().lock();
+            if let Some(leash_data) = leash_data.as_mut() {
+                leash_data.set_holder(holder);
+            } else {
+                *leash_data = Some(LeashData::from_entity(holder));
+            }
+        }
+
+        if self.is_passenger() {
+            self.stop_riding();
+        }
+        if let Some(old_holder) = old_holder
+            && old_holder.id() != holder.id()
+        {
+            old_holder.notify_leashee_removed(self.as_entity_event_source());
+        }
+        true
+    }
+
+    fn tick_leash(&self) {
+        if let Some(holder) = self.leash_holder() {
+            if !self.can_interact_with_level() || !holder.can_interact_with_level() {
+                if let Some(world) = self.level()
+                    && world.get_game_rule(&ENTITY_DROPS)
+                {
+                    self.drop_leash();
+                } else {
+                    self.remove_leash();
+                }
+                return;
+            }
+
+            let distance_to = self.leash_distance_to(holder.as_ref());
+            self.when_leashed_to(holder.as_ref());
+            let angular_momentum_before_distance_action = self.leash_angular_momentum();
+            if distance_to > self.leash_snap_distance() {
+                if let Some(world) = self.level() {
+                    world.play_sound_at(
+                        &sound_events::ITEM_LEAD_BREAK,
+                        SoundSource::Neutral,
+                        holder.position(),
+                        1.0,
+                        1.0,
+                        None,
+                    );
+                }
+                self.leash_too_far_behaviour();
+            } else if distance_to
+                > self.leash_elastic_distance()
+                    - f64::from(holder.base().dimensions().width)
+                    - f64::from(self.base().dimensions().width)
+                && self.check_elastic_interactions(holder.as_ref())
+            {
+                self.on_elastic_leash_pull();
+            } else {
+                self.close_range_leash_behaviour(holder.as_ref());
+            }
+            if !self.apply_leash_angular_momentum()
+                && let Some(angular_momentum) = angular_momentum_before_distance_action
+            {
+                self.rotate_by_leash_angular_momentum(angular_momentum);
+            }
+            return;
+        }
+
+        let Some(attachment) = self.leash_attachment() else {
+            return;
+        };
+
+        let Some(world) = self.level() else {
+            return;
+        };
+
+        match attachment {
+            LeashAttachment::Entity(uuid) => {
+                if let Some(holder) = world.get_entity_by_uuid(&uuid) {
+                    let _ = self.set_leashed_to(&holder);
+                    return;
+                }
+
+                if self.tick_count() > DELAYED_LEASH_DROP_TICKS {
+                    let _ = self.spawn_at_location(ItemStack::new(&vanilla_items::LEAD), 0.0);
+                    self.remove_leash_state();
+                }
+            }
+            LeashAttachment::FenceKnot(pos) => {
+                if let Some(holder) = LeashFenceKnotEntity::get_or_create_knot(&world, pos) {
+                    let _ = self.set_leashed_to(&holder);
+                    return;
+                }
+
+                if self.tick_count() > DELAYED_LEASH_DROP_TICKS {
+                    let _ = self.spawn_at_location(ItemStack::new(&vanilla_items::LEAD), 0.0);
+                    self.remove_leash_state();
+                }
+            }
+        }
+    }
+
+    fn drop_leash(&self) {
+        if self.leash_holder().is_none() {
+            return;
+        }
+
+        let holder = self.remove_leash_state();
+        let _ = self.spawn_at_location(ItemStack::new(&vanilla_items::LEAD), 0.0);
+        if let Some(holder) = holder {
+            holder.notify_leashee_removed(self.as_entity_event_source());
+        }
+    }
+
+    fn remove_leash(&self) {
+        if self.leash_holder().is_some()
+            && let Some(holder) = self.remove_leash_state()
+        {
+            holder.notify_leashee_removed(self.as_entity_event_source());
+        }
+    }
+
+    fn remove_leash_state(&self) -> Option<SharedEntity> {
+        self.leash_data()
+            .lock()
+            .take()
+            .and_then(|leash_data| leash_data.holder())
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum LeashAttachment {
+    Entity(Uuid),
+    FenceKnot(BlockPos),
+}
+
+#[derive(Debug, Clone)]
+pub struct LeashData {
+    pub attachment: LeashAttachment,
+    pub holder: Option<WeakEntity>,
+    pub angular_momentum: f64,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub(super) struct LeashWrench {
+    pub(super) force: DVec3,
+    pub(super) torque: f64,
+}
+
+impl LeashWrench {
+    pub const fn new(force: DVec3, torque: f64) -> Self {
+        Self { force, torque }
+    }
+}
+
+impl LeashData {
+    pub(crate) fn from_entity(holder: &SharedEntity) -> Self {
+        let attachment = holder.downcast_ref::<LeashFenceKnotEntity>().map_or_else(
+            || LeashAttachment::Entity(holder.uuid()),
+            |knot| LeashAttachment::FenceKnot(knot.block_pos()),
+        );
+        Self {
+            attachment,
+            holder: Some(Arc::downgrade(holder)),
+            angular_momentum: 0.0,
+        }
+    }
+
+    pub(crate) const fn from_delayed_attachment(attachment: LeashAttachment) -> Self {
+        Self {
+            attachment,
+            holder: None,
+            angular_momentum: 0.0,
+        }
+    }
+
+    pub(super) fn holder(&self) -> Option<SharedEntity> {
+        self.holder.as_ref().and_then(WeakEntity::upgrade)
+    }
+
+    pub(super) fn saved_attachment(&self) -> LeashAttachment {
+        self.holder().map_or(self.attachment, |holder| {
+            holder.downcast_ref::<LeashFenceKnotEntity>().map_or_else(
+                || LeashAttachment::Entity(holder.uuid()),
+                |knot| LeashAttachment::FenceKnot(knot.block_pos()),
+            )
+        })
+    }
+
+    pub(super) fn set_holder(&mut self, holder: &SharedEntity) {
+        self.attachment = holder.downcast_ref::<LeashFenceKnotEntity>().map_or_else(
+            || LeashAttachment::Entity(holder.uuid()),
+            |knot| LeashAttachment::FenceKnot(knot.block_pos()),
+        );
+        self.holder = Some(Arc::downgrade(holder));
+        self.angular_momentum = 0.0;
+    }
+
+    pub(super) fn save(&self, nbt: &mut NbtCompound) {
+        match self.saved_attachment() {
+            LeashAttachment::Entity(uuid) => {
+                let mut leash = NbtCompound::new();
+                leash.insert("UUID", NbtTag::IntArray(uuid.to_int_array().to_vec()));
+                nbt.insert("leash", NbtTag::Compound(leash));
+            }
+            LeashAttachment::FenceKnot(pos) => {
+                nbt.insert("leash", NbtTag::IntArray(vec![pos.x(), pos.y(), pos.z()]));
+            }
+        }
+    }
+
+    pub(super) fn load(nbt: BorrowedNbtCompoundView<'_, '_>) -> Option<Self> {
+        if let Some(leash) = nbt.compound("leash")
+            && let Some(uuid_array) = leash.int_array("UUID")
+            && let Some(uuid) = Uuid::from_int_array(&uuid_array)
+        {
+            return Some(Self::from_delayed_attachment(LeashAttachment::Entity(uuid)));
+        }
+
+        nbt.int_array("leash")
+            .filter(|position| position.len() == 3)
+            .map(|position| {
+                Self::from_delayed_attachment(LeashAttachment::FenceKnot(BlockPos::new(
+                    position[0],
+                    position[1],
+                    position[2],
+                )))
+            })
+    }
+}
+
+pub(super) fn leash_dimensions(entity: &dyn Entity) -> DVec3 {
+    let dimensions = entity.base().dimensions();
+    DVec3::new(
+        f64::from(dimensions.width),
+        f64::from(dimensions.height),
+        f64::from(dimensions.width),
+    )
+}
+
+pub(super) fn leash_bounding_box_center(entity: &dyn Entity) -> DVec3 {
+    let bounding_box = entity.bounding_box();
+    DVec3::new(
+        f64::midpoint(bounding_box.min_x(), bounding_box.max_x()),
+        f64::midpoint(bounding_box.min_y(), bounding_box.max_y()),
+        f64::midpoint(bounding_box.min_z(), bounding_box.max_z()),
+    )
+}
+
+pub(super) fn leash_holder_movement(entity: &dyn Entity) -> DVec3 {
+    if entity.as_mob().is_some_and(Mob::is_no_ai) {
+        return DVec3::ZERO;
+    }
+
+    entity.known_movement()
+}
+
+pub(super) fn rotate_y(vector: DVec3, radians: f32) -> DVec3 {
+    let cos = f64::from(radians.cos());
+    let sin = f64::from(radians.sin());
+    DVec3::new(
+        vector.x * cos + vector.z * sin,
+        vector.y,
+        vector.z * cos - vector.x * sin,
+    )
+}
+
+pub(super) fn axis_specific_leash_elasticity(force: DVec3) -> DVec3 {
+    force * LEASH_AXIS_SPECIFIC_ELASTICITY
+}
+
+pub(super) fn compute_elastic_interaction(
+    entity: &dyn Entity,
+    holder: &dyn Entity,
+    slack_distance: f64,
+) -> Option<LeashWrench> {
+    let entity_y_rot = entity.rotation().0 * PI / 180.0;
+    let entity_attach_vector = rotate_y(
+        ENTITY_LEASH_ATTACHMENT_POINT * leash_dimensions(entity),
+        -entity_y_rot,
+    );
+    let entity_attach_pos = entity.position() + entity_attach_vector;
+
+    let holder_y_rot = holder.rotation().0 * PI / 180.0;
+    let holder_attach_vector = rotate_y(
+        LEASHER_ATTACHMENT_POINT * leash_dimensions(holder),
+        -holder_y_rot,
+    );
+    let holder_attach_pos = holder.position() + holder_attach_vector;
+
+    compute_dampened_spring_interaction(
+        holder_attach_pos,
+        entity_attach_pos,
+        slack_distance,
+        leash_holder_movement(entity),
+        entity_attach_vector,
+    )
+}
+
+pub(super) fn compute_dampened_spring_interaction(
+    pivot_point: DVec3,
+    object_position: DVec3,
+    spring_slack: f64,
+    object_motion: DVec3,
+    lever_arm: DVec3,
+) -> Option<LeashWrench> {
+    let distance = object_position.distance(pivot_point);
+    if distance < spring_slack {
+        return None;
+    }
+
+    let mut displacement = (pivot_point - object_position).normalize() * (distance - spring_slack);
+    let torque = torque_from_force(lever_arm, displacement);
+    if object_motion.dot(displacement) >= 0.0 {
+        displacement *= 1.0 - LEASH_SPRING_DAMPENING;
+    }
+
+    Some(LeashWrench::new(displacement, torque))
+}
+
+pub(super) fn torque_from_force(lever_arm: DVec3, force: DVec3) -> f64 {
+    lever_arm.z * force.x - lever_arm.x * force.z
+}
